change(electron): changes writing generated files usability
Now all project files are written at once.

This is more convenient and is needed because of the game behavior.

If you attempt to load an audio game data file that contains information about one CMloArchetypeDef that is inside a #typ file with multiple others CMloArchetypeDef, the data file is not loaded.